requirements.txt
Flask==3.0.0
qrcode[pil]==7.4.2
openpyxl==3.1.2
pycryptodome==3.19.0
Pillow>=10.0.0
Werkzeug==3.0.1
